+++ jscomp/core/js_exp_make.mli
@@ -111,6 +111,8 @@ val uint32 : ?comment:string -> int32 -> t
 
 val small_int : int -> t
 
+val bigint : ?comment:string -> bool -> string -> t
+
 val float : ?comment:string -> string -> t
 
 (* val empty_string_literal : t  *)
@@ -120,6 +122,8 @@ val zero_int_literal : t
 (* val one_int_literal : t *)
 val zero_float_lit : t
 (* val obj_int_tag_literal : t *)
+
+val zero_bigint_literal : t
 
 val is_out : ?comment:string -> t -> t -> t
 (** [is_out e range] is equivalent to [e > range or e <0]
@@ -271,6 +275,14 @@ val bool_comp : Lam_compat.comparison -> ?comment:string -> t -> t -> t
 val string_comp : Js_op.binop -> ?comment:string -> t -> t -> t
 
 val float_comp : Lam_compat.comparison -> ?comment:string -> t -> t -> t
+
+val bigint_op : ?comment: string -> Js_op.binop -> t -> t -> t
+
+val bigint_comp : Lam_compat.comparison -> ?comment:string -> t -> t -> t
+
+val bigint_div : checked:bool -> ?comment:string -> t -> t -> t
+
+val bigint_mod : checked:bool -> ?comment:string -> t -> t -> t
 
 val js_comp : Lam_compat.comparison -> ?comment:string -> t -> t -> t