+++ jscomp/core/lam_primitive.ml
@@ -80,10 +80,24 @@ type t =
   | Psubfloat
   | Pmulfloat
   | Pdivfloat
+  (* Bigint operations *)
+  | Pnegbigint
+  | Paddbigint
+  | Psubbigint
+  | Pmulbigint
+  | Pdivbigint
+  | Pmodbigint
+  | Ppowbigint
+  | Pandbigint
+  | Porbigint
+  | Pxorbigint
+  | Plslbigint
+  | Pasrbigint
   | Pintcomp of Lam_compat.comparison
   | Pfloatcomp of Lam_compat.comparison
   | Pjscomp of Lam_compat.comparison
   | Pint64comp of Lam_compat.comparison
+  | Pbigintcomp of Lam_compat.comparison
   | Pjs_apply (*[f;arg0;arg1; arg2; ... argN]*)
   | Pjs_runtime_apply (* [f; [...]] *)
   (* String operations *)
@@ -201,6 +215,18 @@ let eq_primitive_approx (lhs : t) (rhs : t) =
   | Psubfloat -> rhs = Psubfloat
   | Pmulfloat -> rhs = Pmulfloat
   | Pdivfloat -> rhs = Pdivfloat
+  | Pnegbigint -> rhs = Pnegbigint
+  | Paddbigint -> rhs = Paddbigint
+  | Psubbigint -> rhs = Psubbigint
+  | Pmulbigint -> rhs = Pmulbigint
+  | Pdivbigint -> rhs = Pdivbigint
+  | Pmodbigint -> rhs = Pmodbigint
+  | Ppowbigint -> rhs = Ppowbigint
+  | Pandbigint -> rhs = Pandbigint
+  | Porbigint -> rhs = Porbigint
+  | Pxorbigint -> rhs = Pxorbigint
+  | Plslbigint -> rhs = Plslbigint
+  | Pasrbigint -> rhs = Pasrbigint
   | Pjs_apply -> rhs = Pjs_apply
   | Pjs_runtime_apply -> rhs = Pjs_runtime_apply
   | Pstringlength -> rhs = Pstringlength
@@ -267,6 +293,11 @@ let eq_primitive_approx (lhs : t) (rhs : t) =
   | Pfloatcomp comparison -> (
       match rhs with
       | Pfloatcomp comparison1 ->
+          Lam_compat.eq_comparison comparison comparison1
+      | _ -> false)
+  | Pbigintcomp comparison -> (
+      match rhs with
+      | Pbigintcomp comparison1 ->
           Lam_compat.eq_comparison comparison comparison1
       | _ -> false)
   | Pjscomp comparison -> (