+++ jscomp/frontend/lam_constant.ml
@@ -50,6 +50,7 @@ type t =
   | Const_string of {s: string; unicode: bool}
   | Const_float of string
   | Const_int64 of int64
+  | Const_bigint of bool * string
   | Const_pointer of string
   | Const_block of int * Lambda.tag_info * t list
   | Const_float_array of string list
@@ -85,6 +86,10 @@ let rec eq_approx (x : t) (y : t) =
   | Const_int64 ix -> (
     match y with
     | Const_int64 iy -> ix = iy
+    | _ -> false)
+  | Const_bigint (sx, ix) -> (
+    match y with
+    | Const_bigint (sy, iy) -> sx = sy && ix = iy
     | _ -> false)
   | Const_pointer ix -> (
     match y with