+++ jscomp/ml/ast_payload.mli
@@ -41,6 +41,8 @@ val is_single_int : t -> int option
 
 val is_single_float : t -> string option
 
+val is_single_bigint : t -> string option
+
 val is_single_bool : t -> bool option
 
 val is_single_ident : t -> Longident.t option