+++ jscomp/ml/asttypes.ml
@@ -22,7 +22,7 @@ type constant =
   | Const_float of string
   | Const_int32 of int32
   | Const_int64 of int64
-  | Const_nativeint of nativeint
+  | Const_bigint of bool * string
 
 type rec_flag = Nonrecursive | Recursive