+++ jscomp/ml/typecore.ml
@@ -239,8 +239,8 @@ let type_constant = function
   | Const_string _ -> instance_def Predef.type_string
   | Const_float _ -> instance_def Predef.type_float
   | Const_int64 _ -> instance_def Predef.type_int64
-  | Const_int32 _ 
-  | Const_nativeint _ ->  assert false
+  | Const_bigint _ -> instance_def Predef.type_bigint
+  | Const_int32 _ ->  assert false
 
 let constant : Parsetree.constant -> (Asttypes.constant, error) result =
   function
@@ -260,10 +260,8 @@ let constant : Parsetree.constant -> (Asttypes.constant, error) result =
        with Failure _ -> Error (Literal_overflow "int64")
      end
   | Pconst_integer (i,Some 'n') ->
-     begin
-       try Ok (Const_nativeint (Misc.Int_literal_converter.nativeint i))
-       with Failure _ -> Error (Literal_overflow "nativeint")
-     end
+    let sign, i = Bigint_utils.parse_bigint i in
+    Ok (Const_bigint (sign, i))
   | Pconst_integer (i,Some c) -> Error (Unknown_literal (i, c))
   | Pconst_char c -> Ok (Const_char c)
   | Pconst_string (s,d) -> Ok (Const_string (s,d))